Rename ServicesSpecialitiesMap to navLinksHelper
